Move dispatch to component.ts
