+++ src/index.ts
@@ -1,4 +1,3 @@
-export { defineComponent, mountComponent } from './component'
-export { debug } from './dispatch'
+export { defineComponent, mountComponent, debug } from './component'
 export * from './jsxFactory'
 export * from './contract'
